+++ src/main/java/nextstep/auth/authentication/filter/checking/AuthenticationCheckingFilter.java
@@ -1,31 +1,20 @@
 package nextstep.auth.authentication.filter.checking;
 
-import nextstep.auth.authentication.AuthenticationToken;
+import nextstep.auth.authentication.filter.AuthenticationFilter;
 import nextstep.auth.context.Authentication;
 import nextstep.auth.context.SecurityContextHolder;
-import org.springframework.web.servlet.HandlerInterceptor;
 
-import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
 
-public abstract class AuthenticationCheckingFilter implements HandlerInterceptor {
+public abstract class AuthenticationCheckingFilter extends AuthenticationFilter {
 
     @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-        try {
-            var authenticationToken = convert(request);
-            var authentication = authenticate(authenticationToken);
-
-            SecurityContextHolder.getContext().setAuthentication(authentication);
-
-        } catch (Exception e) {
-            return true;
-        }
-        return true;
+    protected void processing(Authentication authentication, HttpServletResponse response) throws Exception {
+        SecurityContextHolder.getContext().setAuthentication(authentication);
     }
 
-    protected abstract AuthenticationToken convert(HttpServletRequest request) throws IOException;
-
-    protected abstract Authentication authenticate(AuthenticationToken authenticationToken);
+    @Override
+    protected boolean proceed() {
+        return true;
+    }
 }
